% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnps_ms2}
\alias{gnps_ms2}
\title{Global Natural Product Social Molecular Networking}
\format{
A databaseClass object.
}
\usage{
gnps_ms2
}
\description{
GNPS
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
\keyword{datasets}
